GraphQL tests refactoring
Separate tests in different classes
Replay all POST tests with BodyHandler
