+++ vertx-web-graphql/src/test/java/io/vertx/ext/web/handler/graphql/PostRequestsWithBodyTest.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2019 Red Hat, Inc.
+ *
+ * Red Hat licenses this file to you under the Apache License, version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.vertx.ext.web.handler.graphql;
+
+import io.vertx.ext.web.handler.BodyHandler;
+
+/**
+ * @author Thomas Segismont
+ */
+public class PostRequestsWithBodyTest extends PostRequestsTest {
+
+  @Override
+  public void setUp() throws Exception {
+    super.setUp();
+    router.route("/graphql").order(99).handler(BodyHandler.create());
+  }
+}
